refactor: extract searchable-dropdown.js
